requirements.txt
pandas>=1.5
networkx>=3.0
numpy>=1.23
scipy>=1.9
matplotlib>=3.6
seaborn>=0.12
pyyaml>=6.0
